--- a/runtime/Cargo.toml
+++ b/runtime/Cargo.toml
@@ -30,7 +30,6 @@ itertools = { workspace = true }
 lazy_static = { workspace = true }
 libc = { workspace = true }
 log = { workspace = true }
-lru = { workspace = true }
 lz4 = { workspace = true }
 memmap2 = { workspace = true }
 mockall = { workspace = true }