Moved from https://github.com/solana-labs/solana/pull/35228

#### Problem

`lru::LruCache` requires `write` lock for any action.

#### Summary of Changes

Use `BTreeMap` instead of `LruCache`, write lock would be required only on finalizing slot + `BTreeMap` allows easily to remove the oldest slot.