@tao-stones 2024-03-13
Can you benchmark before/after change for comparison? 

This would run existing bench, if you found these benches are outdated, please feel free to update them 😄 
`./cargo nightly bench --manifest-path runtime/Cargo.toml -- -Z unstable-options bench_process_transactions --ignored`