@fanatid 2024-03-14
master 151675b5ca
```
running 2 tests
test bench_process_transactions_multiple_slots ... bench:  13,967,987 ns/iter (+/- 1,213,641)
test bench_process_transactions_single_slot    ... bench:   2,059,505 ns/iter (+/- 411,210)
```

PR 30306112d1
```
running 2 tests
test bench_process_transactions_multiple_slots ... bench:  12,907,941 ns/iter (+/- 1,044,988)
test bench_process_transactions_single_slot    ... bench:   1,339,063 ns/iter (+/- 239,231)
```

but I do not think that this display real improvements because we only parse transactions and send results to background update thread
not sure how to benchmark background thread, so I added std output (where cache_lock_time few times less)

master
```
running 1 test
slot_finalize_time 128950us, cache_lock_time 7069us
test bench_process_transactions_single_slot    ... bench:   1,704,378 ns/iter (+/- 186,777)
```

PR
```
running 1 test
slot_finalize_time 119721us, cache_lock_time 1765us
test bench_process_transactions_single_slot    ... bench:   1,592,675 ns/iter (+/- 249,831)
```

diff:
```diff
diff --git a/runtime/benches/prioritization_fee_cache.rs b/runtime/benches/prioritization_fee_cache.rs
index 8c6bf1fe0a..f97691625d 100644
--- a/runtime/benches/prioritization_fee_cache.rs
+++ b/runtime/benches/prioritization_fee_cache.rs
@@ -44,22 +44,29 @@ fn build_sanitized_transaction(
 fn bench_process_transactions_single_slot(bencher: &mut Bencher) {
     let prioritization_fee_cache = PrioritizationFeeCache::default();
 
-    let bank = Arc::new(Bank::default_for_tests());
+    let GenesisConfigInfo { genesis_config, .. } = create_genesis_config(10_000);
+    let bank0 = Bank::new_for_benches(&genesis_config);
+    let bank_forks = BankForks::new_rw_arc(bank0);
+    let bank = bank_forks.read().unwrap().working_bank();
+    let collector = solana_sdk::pubkey::new_rand();
+    let mut n = 0;
+    bencher.iter(move || {
+        n += 1;
+        let bank = Bank::new_from_parent(bank.clone(), &collector, n);
 
-    // build test transactions
-    let transactions: Vec<_> = (0..5000)
-        .map(|n| {
-            let compute_unit_price = n % 7;
-            build_sanitized_transaction(
-                compute_unit_price,
-                &Pubkey::new_unique(),
-                &Pubkey::new_unique(),
-            )
-        })
-        .collect();
+        let transactions = (0..500)
+            .map(|n| {
+                let compute_unit_price = n % 7;
+                build_sanitized_transaction(
+                    compute_unit_price,
+                    &Pubkey::new_unique(),
+                    &Pubkey::new_unique(),
+                )
+            })
+            .collect::<Vec<_>>();
 
-    bencher.iter(|| {
         prioritization_fee_cache.update(&bank, transactions.iter());
+        prioritization_fee_cache.finalize_priority_fee(bank.slot(), bank.bank_id());
     });
 }
 
diff --git a/runtime/src/prioritization_fee_cache.rs b/runtime/src/prioritization_fee_cache.rs
index 0490f59445..176a7ba405 100644
--- a/runtime/src/prioritization_fee_cache.rs
+++ b/runtime/src/prioritization_fee_cache.rs
@@ -44,6 +44,9 @@ struct PrioritizationFeeCacheMetrics {
 
     // Accumulated time spent on finalizing block prioritization fees.
     total_block_finalize_elapsed_us: AtomicU64,
+
+    finalized: AtomicU64,
+    cache: AtomicU64,
 }
 
 impl PrioritizationFeeCacheMetrics {
@@ -65,6 +68,7 @@ impl PrioritizationFeeCacheMetrics {
     fn accumulate_total_cache_lock_elapsed_us(&self, val: u64) {
         self.total_cache_lock_elapsed_us
             .fetch_add(val, Ordering::Relaxed);
+        self.cache.fetch_add(val, Ordering::Relaxed);
     }
 
     fn accumulate_total_entry_update_elapsed_us(&self, val: u64) {
@@ -75,6 +79,7 @@ impl PrioritizationFeeCacheMetrics {
     fn accumulate_total_block_finalize_elapsed_us(&self, val: u64) {
         self.total_block_finalize_elapsed_us
             .fetch_add(val, Ordering::Relaxed);
+        self.finalized.fetch_add(val, Ordering::Relaxed);
     }
 
     fn report(&self, slot: Slot) {
@@ -161,6 +166,7 @@ impl Drop for PrioritizationFeeCache {
             .unwrap()
             .join()
             .expect("Prioritization fee cache servicing thread failed to join");
+        println!("slot_finalize_time {}us, cache_lock_time {}us", self.metrics.finalized.load(Ordering::Relaxed), self.metrics.cache.load(Ordering::Relaxed));
     }
 }
```