@tao-stones 2024-03-14
> so I added std output (where cache_lock_time few times less)

Great to see significant decrease on cache_lock_time! 

Do you mind to add `entry_update_time` to std output alone with `cache_lock_time` and `slot_finalize_time` to get a better picture? Thank you.