@fanatid 2024-03-14
master:
```
running 1 test
cache 7620us, update: 158405us, finalize 128463us
test bench_process_transactions_single_slot    ... bench:   1,747,303 ns/iter (+/- 323,183)
```

PR
```
running 1 test
cache 1527us, update: 392736us, finalize 116258us
test bench_process_transactions_single_slot    ... bench:   1,533,384 ns/iter (+/- 213,986)
```

PR after [bf3cae6](https://github.com/anza-xyz/agave/pull/30/commits/bf3cae6a81eac6e1904748015c0c5de85b342ddf) (update per tx instead of per bank)
```
running 1 test
cache 1785us, update: 149256us, finalize 107472us
test bench_process_transactions_single_slot    ... bench:   1,674,978 ns/iter (+/- 414,933)
```
but honestly `entry_update_time` is very unstable here, I had seen that numbers changed for master and updated PR too much, I had seen 2 times difference:
```
running 1 test
cache 1785us, update: 149256us, finalize 107472us
test bench_process_transactions_single_slot    ... bench:   1,674,978 ns/iter (+/- 414,933)
running 1 test
cache 3929us, update: 318381us, finalize 243182us
test bench_process_transactions_single_slot    ... bench:   1,690,419 ns/iter (+/- 191,294)
```

not sure would be good to send an update per tx or collect + send per bank