@tao-stones 2024-03-15
> this seems to be making several independent changes simultaneously. we generally reject these in favor of one pr per change. i'd highly recommend breaking it up

Yea, agree it'd be better (at least easier to review) if can break into separate PRs, perhaps:
- replace `lruCache` with `BtreeMap` with manual capacity control
- Introducing `unfinalized` cache at receiver thread, 
- batch updates