@fanatid 2024-03-16
`lruCache` and `unfinalized` are related to each other, code would need to be adjusted after each PR again.
While it's possible and I moved `unfinalized` to #272 but doesn't make sense at all, I already split getRPF changes for 2 PRs (this and #217)

The batch update was removed after the benchmark.

`BTreeMap::retain` doesn't make sense, better switch to HashMap then. BTreeMap was used because we iterate in order and can break earlier, HashMap would be faster. I had made a change in #272.