@CriesofCarrots 2024-03-16
> `BTreeMap::retain` doesn't make sense, better switch to HashMap then. BTreeMap was used because we iterate in order and can break earlier, HashMap would be faster. I had made a change in #272.

Did you read the entirety of my comment? `split_off` would take advantage of the ordering and not require visiting every element.